app/requirements.txt
ffmpeg-python
flask
openai-whisper
setuptools-rust
srt
